alerts/historical/requirements.txt
asyncio==3.4.3
numpy==1.21.3
openpyxl==3.0.9
pandas==1.3.4
python-dotenv==0.19.1
simplejson 3.17.5
watchdog==2.1.6